Moved DiagrammedAssertionsMacro.scala into org.scalatest.diagrams package.
