+++ dotty/diagrams/src/main/scala/org/scalatest/diagrams/DiagrammedAssertionsMacro.scala
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
+package org.scalatest.diagrams
 
-import org.scalactic._
 import scala.quoted._
-import org.scalatest.diagrams._
+import org.scalactic._
+import org.scalatest._
 
 object DiagrammedAssertionsMacro {
   /**
